feat: Added match insertion SQL.
Can now insert into both the match and match detail tables simultaneously. Had to add a staging table ID to make this smooth. Reason being: there's not a relationship between the match and match detail ID at the time the match ID is created. The staging table's ID is an intermediary for the time between inserting into the match table and then populating the match detail table.