feat: Add API data cleaning function.
Adds a simple function for turning the 'values'
portion of the dictionary return by Google's
API into a dictionary consisting of column headers
as keys and a list of strings as values (rows).
Adds a test for this function too.

This format will make the data easier to manipulate.
E.g., can be fed directly into a DataFrame later
if needed.

Regarding the test, it's not exactly foolproof,
but it captures if the rough structure of the output
ever messes up.