fix: Change scopes data type.
This thing needs to be a list of strings, as it's
acceptable to pass multiple scopes in and
the Google Cloud API will interpret it correctly.
Only thing to be wary of is your environment
variables correctly being read in as a list of
strings, which I accomplish already with the ast
library in main.py.