feat: Add Postgres interaction.
This commit adds a module for interacting with
the Supabase API, which is the service that
manages my Postgres instance. This is useful
not only for loading staging data into my
database, but also executing stored procedures
and manipulating other objects without the use
of a brute for ORM solution/ other access
management method.

This commit also comes with a related data
cleaning function and test because Supabase
requires data in a certain way.

Renamed 'test_etl.p' to 'test_data_cleaning.py'
so it aligns with its 'src' counterpart.