refactor: Facilitate compatibility with DuckDB.
Instead of passing in the raw JSON and getting
a giant inferred column with huge lists interpreted
as one big VARCHAR column and having to fix that
with goofy SQL, it's easier if the JSON is
already structured as col: row values.

The reason I bake a cleaning function into the
GoogleSheet class is to not have to pass a bunch
of variables into it. The function's only use is
in the context of cleaning the API data, so it
should be more efficient to just pop the sheet
attributes I'm already asking for right into the
cleaning process.